no functional edit button

+++ src/main/java/com/codeup/adlister/controllers/DeleteAdServlet.java
@@ -17,13 +17,18 @@ public class DeleteAdServlet extends HttpServlet {
             response.sendRedirect("/login");
             return;
         }
-        long id = Long.parseLong(request.getParameter("ads"));
-        DaoFactory.getAdsDao().delete(id);
+        if (request.getParameter("id") != null) {
+            long id = Long.parseLong(request.getParameter("id"));
+            DaoFactory.getAdsDao().delete(id);
+//            response.sendRedirect("/ads");
+//        long id = Long.parseLong(request.getParameter("ad"));
+//        DaoFactory.getAdsDao().delete(id);
 
-        // Using this method if keeping the delete.jsp //
-//        request.setAttribute("ads", DaoFactory.getAdsDao().all());
+            // Using this method if keeping the delete.jsp //
+            request.setAttribute("ads", DaoFactory.getAdsDao().all());
 
-        response.sendRedirect("/profile");
+            response.sendRedirect("/profile");
+        }
+
     }
-
 }
